Fred pointed out that the pyngl_version.py file, which gets constructed by setup.py, was being left in the directory. This can cause problems if you import PyNGL from this directory, because it will pick up this version file, and not the one from the site-packages directory.
git-svn-id: https://subversion.ucar.edu/ncldev/pyngl/trunk@9033 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ setup.py
@@ -371,7 +371,8 @@ for array_module in array_modules:
 # Create version file that contains version and array module info.
 #
 #----------------------------------------------------------------------
-  os.system("/bin/rm -rf " + pyngl_vfile)
+  if os.path.exists(pyngl_vfile):
+    os.system("/bin/rm -rf " + pyngl_vfile)
 
   pyngl_version = open('version','r').readlines()[0].strip('\n')
 
@@ -477,6 +478,9 @@ for array_module in array_modules:
       )
 
 #
-# Cleanup: remove the Scripts directory.
+# Cleanup: remove the Scripts directory and pyngl_version.py file.
 #
 os.system("/bin/rm -rf " + pynglex_dir)
+
+if os.path.exists(pyngl_vfile):
+  os.system("/bin/rm -rf " + pyngl_vfile)
